+++ Cauldron/Directory.Build.props
@@ -9,7 +9,7 @@
     <LangVersion>latest</LangVersion>
 
     <!-- Version Information -->
-    <Version>0.16.0</Version>
+    <Version>0.18.0</Version>
     <AssemblyVersion>$(Version)</AssemblyVersion>
     <FileVersion>$(Version)</FileVersion>
 
@@ -27,7 +27,9 @@
     
     <!-- Common Description -->
     <Description>
-        Cauldron is a modular web layer that talks with the Morgana conversational AI platform.
+      Cauldron is a modular web layer that talks with the Morgana conversational AI framework.
+      Built on cutting-edge .NET 10 and leveraging the Blazor Server + SignalR development platform,
+      Cauldron is the official Morgana's frontend.
     </Description>
 
     <!-- Disable packaging by default (enable per-project as needed) -->
