+++ Morgana/Directory.Build.props
@@ -9,7 +9,7 @@
     <LangVersion>latest</LangVersion>
 
     <!-- Version Information -->
-    <Version>0.16.0</Version>
+    <Version>0.18.0</Version>
     <AssemblyVersion>$(Version)</AssemblyVersion>
     <FileVersion>$(Version)</FileVersion>
