+++ Morgana/Morgana.Examples/Agents/ContractAgent.cs
@@ -4,30 +4,12 @@ using Morgana.Framework.Adapters;
 using Morgana.Framework.Attributes;
 using Morgana.Framework.Interfaces;
 
-namespace Morgana.Example.Agents;
+namespace Morgana.Examples.Agents;
 
-/// <summary>
-/// <para>
-/// Example agent demonstrating MonkeyMCP server integration.
-/// This agent can retrieve information about monkeys from a public MCP server.
-/// </para>
-/// <para>
-/// Available tools:
-/// - get_all_monkeys: Returns list of all monkeys
-/// - find_monkey_by_name: Finds specific monkey by name
-/// </para>
-/// <para>
-/// Test queries:
-/// - "Get me a list of all monkeys"
-/// - "Find information about the Baboon"
-/// - "Tell me about monkeys from Africa"
-/// </para>
-/// </summary>
-[HandlesIntent("monkeys")]
-[UsesMCPServers("MonkeyMCP")]
-public class MonkeyAgent : MorganaAgent
+[HandlesIntent("contract")]
+public class ContractAgent : MorganaAgent
 {
-    public MonkeyAgent(
+    public ContractAgent(
         string conversationId,
         ILLMService llmService,
         IPromptResolverService promptResolverService,
